dom/base/test/test_intersectionobserver_no_root.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1313927</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="onLoad()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1313927">Mozilla Bug 1313927</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

    function runTest() {
        var win;

        window.onmessage = function (e) {
            win.close();
            is(e.data, true);
            SimpleTest.finish();
        };

        win = window.open("intersectionobserver_window.html");
    }

    function onLoad() {
        SpecialPowers.pushPrefEnv({"set": [["dom.IntersectionObserver.enabled", true]]}, runTest);
    }

    SimpleTest.waitForExplicitFinish();

</script>
</pre>
<div id="log">
</div>
</body>
</html>
